\name{IDSL.NPA_referenceMSPgenerator}
\alias{IDSL.NPA_referenceMSPgenerator}
\title{
IDSL.NPA Reference MSP Generator
}
\description{
This function creates reference standard .msp files.
}
\usage{
IDSL.NPA_referenceMSPgenerator(NPA_peaklist, refNPAtable, selectedPeaks_IDref)
}
\arguments{
  \item{NPA_peaklist}{
A dataframe peaklist of co-detected peaks
}
  \item{refNPAtable}{
reference NPA table
}
  \item{selectedPeaks_IDref}{
selectedPeaks_IDref
}
}
\value{
A string of standard .msp file
}
