\name{IDSL.NPA_workflow}
\alias{IDSL.NPA_workflow}
\title{
IDSL.NPA workflow
}
\description{
This function executes the NPA workflow.
}
\usage{
IDSL.NPA_workflow(spreadsheet)
}
\arguments{
  \item{spreadsheet}{
NPA spreadsheet
}
}
\value{
This function organizes the IDSL.NPA file processing for better performance using the template spreadsheet.
}
