\name{NPA_xlsxAnalyzer}
\alias{NPA_xlsxAnalyzer}
\title{
xlsx Analyzer for NPA analysis
}
\description{
This function processes the spreadsheet of the NPA spreadsheet tab to ensure the parameter inputs are in agreement with requirements of the NPA analysis.
}
\usage{
NPA_xlsxAnalyzer(spreadsheet)
}
\arguments{
  \item{spreadsheet}{
NPA spreadsheet tab
}
}
\value{
parameters to feed the `NPA_workflow` function.
}
